+++ terraform/deploy_vars.sh
@@ -1,9 +1,7 @@
 #! /bin/bash -xu
-
-export CALCLOUD_VER="v0.4.40"
-export CALDP_VER="v0.2.22"
-export CAL_BASE_IMAGE="stsci/hst-pipeline:CALDP_20230613_CAL_final"
-
+export CALCLOUD_VER="v0.4.41"
+export CALDP_VER="v0.2.23"
+export CAL_BASE_IMAGE="stsci/hst-pipeline:CALDP_20230810_CAL_final"
 export BASE_IMAGE_TAG=`cut -d ":" -f2- <<< ${CAL_BASE_IMAGE} `
 
 export COMMON_IMAGE_TAG="CALCLOUD_${CALCLOUD_VER}-CALDP_${CALDP_VER}-BASE_${BASE_IMAGE_TAG}"
@@ -15,6 +13,10 @@ export CALCLOUD_BUILD_DIR=${CALCLOUD_BUILD_DIR:-""}
 export CALDP_BUILD_DIR=${CALDP_BUILD_DIR:-""}
 
 export TMP_INSTALL_DIR="/tmp/calcloud_install"
+
+# Set account ID in env variable and for terraform
+export ACCOUNT_ID=`aws sts get-caller-identity --output=text | awk '{ print $1 }'`
+export TF_VAR_account_id=$ACCOUNT_ID
 
 # get a couple of things from AWS ssm
 # the env, i.e. sb,dev,test,prod
@@ -44,10 +46,16 @@ export IMAGE_REPO=`echo $repo_url | cut -d '/' -f2`        # hst-repro
 export CALDP_ECR_TAG="batch-${COMMON_IMAGE_TAG}"
 export PREDICT_ECR_TAG="predict-${COMMON_IMAGE_TAG}"
 export TRAINING_ECR_TAG="training-${COMMON_IMAGE_TAG}"
+export AMIROTATION_ECR_TAG="amirotation-${COMMON_IMAGE_TAG}"
+export AMIROTATION_ECR_TAG_UNSCANNED="amirotation-not-scanned"
+export AMIROTATION_ECR_TAG_LATEST="amirotation-latest"
 
 export CALDP_DOCKER_IMAGE="${repo_url}:${CALDP_ECR_TAG}"
 export PREDICT_DOCKER_IMAGE="${repo_url}:${PREDICT_ECR_TAG}"
 export TRAINING_DOCKER_IMAGE="${repo_url}:${TRAINING_ECR_TAG}"
+export AMIROTATION_DOCKER_IMAGE="${repo_url}:${AMIROTATION_ECR_TAG}"
+export AMIROTATION_DOCKER_IMAGE_UNSCANNED="${repo_url}:${AMIROTATION_ECR_TAG_UNSCANNED}"
+export AMIROTATION_DOCKER_IMAGE_LATEST="${repo_url}:${AMIROTATION_ECR_TAG_LATEST}"
 
 # turn CAL_BASE_IMAGE into CSYS_VER by splitting at the :, splitting again by underscore and keeping the
 # first two fields, and then converting to lowercase