+++ changelog.md
@@ -1,2 +1,6 @@
-- default base docker image set to CALDP_20230613_CAL_final
-- default crds update to hst_1077.pmap
+- Created the calcloud-ami-rotation CodeBuild project to take over running
+  the biweekly AMI rotation script previously run by the
+  calcloud-env-AmiRotation Lambda
+- Replaced the deprecated sklearn==0.0 package with scikit-learn==1.0.2
+- default base docker image set to CALDP_cosandpin_CAL_rc1
+- default crds update to hst_1089.pmap